+++ cli/src/main/kotlin/ls/kafka/backup/cli/commands/BackupCli.kt
@@ -1,6 +1,0 @@
-package ls.kafka.backup.cli.commands
-
-import picocli.CommandLine
-
-@CommandLine.Command(subcommands = [Restore::class, Repair::class])
-class BackupCli