+++ CHANGES.md
@@ -5,6 +5,8 @@
 ### Feature
 
 - Breaking change: `readAll()` in `RecordStreamReader` returns a cold `Flow` instead of a `List`.
+- bundle cli `kbackup` as command into docker image
+- require `kbackup` explicitly as first argument
 
 ## 2023-09-13 / 0.1.0